Move `device_info` to `bin-shared`
