+++ rust/headless-client/Cargo.toml
@@ -31,7 +31,6 @@ secrecy = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 serde_variant = { workspace = true }
-smbios-lib = { workspace = true }
 strum = { workspace = true }
 thiserror = { workspace = true }
 # This actually relies on many other features in Tokio, so this will probably
